Add files via upload
added dicestruct to use for diceroller class.  Do see demo of use, included in this same file is the game ("Fortune Teller") I was creating and wanted something to print my dice results.  AI started me out, but failed miserably when I wanted a quick reply for code to print the dice all on the same row.  So I created this myself.